.history/src/weaver_20240824032805.py
#!/usr/bin/env python3
"""
AI Response Weaver

This application monitors a specified file for changes, parses AI-generated code blocks,
and creates or updates corresponding files based on the content.

Usage:
    weaver <file_to_monitor> <log_folder>
"""

import os
import sys
import json
import logging
from pathlib import Path
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Set up logging
logging.basicConfig(level=logging.INFO,
                    format='%(asctime)s - %(levelname)s - %(message)s')


class AIResponseWeaver:
    def __init__(self, file_to_monitor: str, log_folder: str):
        self.file_to_monitor = file_to_monitor
        self.log_folder = log_folder
        self.config = self.load_config()

    def load_config(self) -> dict:
        """Load the configuration from config.json"""
        config_path = Path(__file__).parent.parent / 'config' / 'config.json'
        with open(config_path, 'r') as config_file:
            return json.load(config_file)

    def run(self):
        """Main execution method"""
        logging.info(f"Monitoring file: {self.file_to_monitor}")
        logging.info(f"Log folder: {self.log_folder}")
        # TODO: Implement file monitoring and processing logic


def main():
    if len(sys.argv) != 3:
        print(__doc__)
        sys.exit(1)

    file_to_monitor = sys.argv[1]
    log_folder = sys.argv[2]

    weaver = AIResponseWeaver(file_to_monitor, log_folder)
    weaver.run()


if __name__ == "__main__":
    main()
