.history/requirements_20240824032943.txt
watchdog==2.1.9
gitpython==3.1.30
python-dotenv==0.19.2